re-setting up sketch screen
