+++ src/screens/create-task/create-task.container.js
@@ -18,6 +18,7 @@ const mapStateToProps = (state) => {
 const mapDispatchToProps = (dispatch) => {
   return {
     createTask(taskData) {
+      console.log("teste2: " +  taskData.picture_base);
       const callback = (createdTask) => {
         dispatch(taskAdd(createdTask));
