+++ src/screens/index.js
@@ -19,7 +19,7 @@ import EditTaskScreen from './edit-task/edit-task.container';
 import InspirationListScreen from './inspiration-list/inspiration-list.container'
 import InspirationSearchListScreen from './inspiration-search-list/inspiration-search-list.container'
 import ViewInspirationScreen from './view-inspiration/view-inspiration.container'
-import SketchScreen from './sketch/sketch.component';
+import SketchScreen from './sketch/sketch.container';
 
 export {
   CreateUserScreen,
